ward-events: widen EventKindSet to a u64 backing ahead of the 33rd kind
Current main plus this PR reaches exactly 32 event kinds, the full
capacity of the u32 EventKindSet previously backed itself with. #197
(queued to land before this PR per the documented release order) adds a
33rd kind, LaunchAborted, which the u32 backing cannot represent at all:
EventKind::bit()'s 1 << discriminant shift would itself overflow first.

Widened EventKindSet's backing (and EventKind::bit()'s return type) from
u32 to u64, giving 32 kinds of headroom instead of zero. MASK now
computes via checked_shl against a u64::MAX fallback rather than
borrowing headroom from a wider primitive (there isn't a wider one to
borrow from this time).

Wire compatibility: postcard's integer encoding is a plain unsigned
varint with no width tag, so a value that fits in the old u32 backing
serializes identically whether declared u32 or u64. Verified against
real postcard bytes, not asserted from the format's docs:
a_u32_encoded_set_decodes_identically_as_the_widened_u64_type encodes
both u32::MAX and a sparse bit pattern as bare u32s (the exact wire
shape any EventKindSet produced before this change would have had) and
confirms they decode through today's u64-backed Deserialize impl to the
same values, and that re-encoding produces the identical bytes back. No
WIRE_VERSION bump: this is a pure capacity change, not a change to what
any existing bit means.

Rewrote kind_bits_are_dense_and_cover_the_mask's capacity assertion: at
32 kinds the u32 backing was exactly saturated (no unused bit existed to
reject), but with 32 bits of new headroom there is now a first unused
bit (32) that must be rejected as unknown rather than silently accepted
-- the same density property, checked against the new state instead of
the old exhausted one.

crates/ward-events/tests/roundtrip.rs's subscribe_roundtrips proptest
strategy for `kinds` moved from 0u32..(1 << 27) to 0u64..(1 << 27) (same
numeric range) so it still resolves through EventKindSet's TryFrom impl,
now TryFrom<u64> rather than TryFrom<u32>.

+++ crates/ward-events/tests/roundtrip.rs
@@ -221,7 +221,7 @@ proptest! {
     }
 
     #[test]
-    fn subscribe_roundtrips(session in any::<u128>(), from_seq in any::<u64>(), bits in 0u8..128, kinds in 0u32..(1 << 27), notes in any::<bool>()) {
+    fn subscribe_roundtrips(session in any::<u128>(), from_seq in any::<u64>(), bits in 0u8..128, kinds in 0u64..(1 << 27), notes in any::<bool>()) {
         let sub = Subscribe {
             session: SessionId::from_u128(session),
             from_seq,
